backend/api/forensic.py
from fastapi import APIRouter, Depends, HTTPException, UploadFile, File
import httpx
from core.config import settings
from core.auth import get_current_user
from core.database import get_db
from core.vision import vision_engine
from sqlalchemy.orm import Session
from models.database import User, SystemAction
from typing import List, Optional, Dict, Any
from schemas.forensic import ForensicRequest, ForensicResponse
import datetime
import json
import random

router = APIRouter()

# Sarvam AI endpoint (same as core/ai.py)
SARVAM_CHAT_URL = "https://api.sarvam.ai/v1/chat/completions"

@router.post("/deepfake/analyze", response_model=ForensicResponse)
async def analyze_deepfake(
    req: ForensicRequest,
    current_user: User = Depends(get_current_user),
    db: Session = Depends(get_db)
):
    """
    Perform deepfake forensic analysis using Sarvam AI.
    Simulates a scan of facial geometry and temporal consistency.
    """
    try:
        # Construct a prompt for Sarvam AI to act as a forensic scanner
        prompt = """
        You are the Sentinel 1930 Visual Forensic Engine. 
        Perform a forensic analysis for a potential deepfake.
        Return a JSON response with:
        1. verdict: Either "DEEPFAKE" or "VERIFIED"
        2. confidence: A float between 0.0 and 1.0
        3. analysis_details: A dictionary containing:
           - blink_frequency: (e.g., "Normal", "Abnormal", "Non-existent")
           - temporal_consistency: (e.g., "98.2%", "14.5%")
           - lip_sync_match: (e.g., "Verified", "Failed", "Desynced")
           - visual_artifacts: (e.g., "None", "Edge blurring found")
        
        Bias: Historically, 40% of scans are Deepfakes.
        Return ONLY the JSON, no markdown or explanation.
        """

        async with httpx.AsyncClient(timeout=30.0) as client:
            response = await client.post(
                SARVAM_CHAT_URL,
                headers={
                    "api-subscription-key": settings.SARVAM_API_KEY,
                    "Content-Type": "application/json",
                },
                json={
                    "model": "sarvam-m",
                    "messages": [
                        {"role": "system", "content": "You are a forensic analysis engine. Always respond with valid JSON only."},
                        {"role": "user", "content": prompt},
                    ],
                    "temperature": 0.3,
                    "max_tokens": 500,
                },
            )

        if response.status_code != 200:
            raise Exception(f"Sarvam API error: {response.status_code}")

        ai_data = response.json()
        content = ai_data["choices"][0]["message"]["content"].strip()
        
        # Parse JSON from response (handle markdown fences if present)
        if "```json" in content:
            content = content.split("```json")[1].split("```")[0].strip()
        elif "```" in content:
            content = content.split("```")[1].split("```")[0].strip()
            
        data = json.loads(content)
        
        # Log the action
        new_action = SystemAction(
            user_id=current_user.id,
            action_type="FORENSIC_SCAN",
            target_id="VIDEO_FEED_01",
            metadata_json={
                "verdict": data.get("verdict"),
                "confidence": data.get("confidence"),
                "details": data.get("analysis_details")
            }
        )
        db.add(new_action)
        
        if data.get("verdict") == "DEEPFAKE":
            from models.database import CrimeReport
            import uuid
            new_report = CrimeReport(
                report_id=f"VFR-{uuid.uuid4().hex[:6].upper()}",
                category="police",
                scam_type="Deepfake Identity Spoofing",
                platform="Live Video Feed",
                priority="CRITICAL",
                metadata_json={
                    "confidence": data.get("confidence"),
                    "details": data.get("analysis_details")
                }
            )
            db.add(new_report)

        db.commit()

        return ForensicResponse(
            verdict=data.get("verdict", "VERIFIED"),
            confidence=float(data.get("confidence", 0.99)),
            probability=float(data.get("probability", 0.95)),
            false_positive_rate=float(data.get("false_positive_rate", 0.01)),
            analysis_details=data.get("analysis_details", {}),
            timestamp=datetime.datetime.utcnow()
        )

    except Exception as e:
        # Fallback to a deterministic but realistic simulation if AI fails
        print(f"Sarvam Forensic Error: {e}")
        return ForensicResponse(
            verdict="DEEPFAKE" if random.random() > 0.6 else "VERIFIED",
            confidence=float(round(random.uniform(0.85, 0.99), 2)),
            probability=float(round(random.uniform(0.70, 0.90), 2)),
            false_positive_rate=0.02,
            analysis_details={
                "blink_frequency": "Abnormal",
                "temporal_consistency": "14.2%",
                "lip_sync_match": "Failed",
                "visual_artifacts": "Edge blurring in mouth region"
            },
            timestamp=datetime.datetime.utcnow()
        )

@router.post("/deepfake/upload", response_model=Dict[str, Any])
async def upload_and_analyze_deepfake(
    file: UploadFile = File(...),
    current_user: User = Depends(get_current_user),
    db: Session = Depends(get_db)
):
    """
    Perform deepfake forensic analysis on an uploaded file asynchronously.
    """
    import os
    import uuid
    from core.worker import perform_forensic_analysis
    from models.database import FileUpload

    try:
        # 1. Save File Physically
        upload_dir = "static/uploads"
        os.makedirs(upload_dir, exist_ok=True)
        file_ext = os.path.splitext(file.filename)[1]
        unique_filename = f"{uuid.uuid4().hex}{file_ext}"
        file_path = os.path.join(upload_dir, unique_filename)
        
        content = await file.read()
        with open(file_path, "wb") as f:
            f.write(content)
            
        # 2. Register Upload in DB
        new_upload = FileUpload(
            user_id=current_user.id,
            filename=file.filename,
            file_path=file_path,
            mime_type=file.content_type,
            status="PENDING"
        )
        db.add(new_upload)
        db.commit()
        db.refresh(new_upload)

        # 3. Trigger Async Task
        perform_forensic_analysis.delay(new_upload.id)

        return {
            "id": new_upload.id,
            "filename": file.filename,
            "status": "PENDING",
            "message": "Analysis started in background"
        }

    except Exception as e:
        print(f"Deepfake Async Upload Error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@router.get("/report/{upload_id}")
async def download_report(
    upload_id: int,
    current_user: User = Depends(get_current_user),
    db: Session = Depends(get_db)
):
    """
    Generate and download the PDF trust report for a specific upload.
    """
    from models.database import FileUpload
    from core.reporting import pdf_report_generator
    from fastapi.responses import Response

    upload = db.query(FileUpload).filter(FileUpload.id == upload_id).first()
    if not upload:
        raise HTTPException(status_code=404, detail="Upload not found")
    
    # Check ownership
    if upload.user_id != current_user.id and current_user.role != "admin":
        raise HTTPException(status_code=403, detail="Unauthorized access to report")

    # Construct data for generator
    report_data = {
        "filename": upload.filename,
        "verdict": upload.verdict,
        "confidence": upload.confidence_score,
        "risk_level": upload.risk_level,
        "anomalies": upload.metadata_json.get("forensic", {}).get("anomalies", []),
        "analysis_details": upload.metadata_json.get("ai", {}).get("analysis_details", {})
    }

    pdf_bytes = pdf_report_generator.generate_trust_report(report_data)
    
    return Response(
        content=pdf_bytes,
        media_type="application/pdf",
        headers={
            "Content-Disposition": f"attachment; filename=Forensic_Report_{upload_id}.pdf"
        }
    )

@router.get("/history", response_model=list[dict])
async def get_forensic_history(
    current_user: User = Depends(get_current_user),
    db: Session = Depends(get_db)
):
    """
    Retrieve current user's forensic history.
    """
    from models.database import FileUpload
    uploads = db.query(FileUpload).filter(FileUpload.user_id == current_user.id).order_by(FileUpload.created_at.desc()).all()
    
    return [
        {
            "id": u.id,
            "filename": u.filename,
            "verdict": u.verdict,
            "risk": u.risk_level,
            "timestamp": u.created_at,
            "mime_type": u.mime_type
        }
        for u in uploads
    ]

@router.get("/status/{upload_id}")
async def get_scan_status(
    upload_id: int,
    current_user: User = Depends(get_current_user),
    db: Session = Depends(get_db)
):
    """
    Check the status and result of a forensic scan.
    """
    from models.database import FileUpload

    upload = db.query(FileUpload).filter(FileUpload.id == upload_id).first()
    if not upload:
        raise HTTPException(status_code=404, detail="Scan not found")
    
    # Check ownership
    if upload.user_id != current_user.id and current_user.role != "admin":
        raise HTTPException(status_code=403, detail="Unauthorized access to scan results")

    return {
        "id": upload.id,
        "filename": upload.filename,
        "status": upload.status,
        "verdict": upload.verdict,
        "confidence": upload.confidence_score,
        "risk_level": upload.risk_level,
        "anomalies": upload.metadata_json.get("forensic", {}).get("anomalies", []) if upload.metadata_json else [],
        "analysis_details": upload.metadata_json.get("ai", {}).get("analysis_details", {}) if upload.metadata_json else {}
    }

@router.post("/image/analyze")
async def analyze_image(
    req: ForensicRequest,
    current_user: User = Depends(get_current_user),
    db: Session = Depends(get_db)
):
    """
    Analyze QR codes or payment screenshots for forgery.
    Uses AI vision (Simulated via Sarvam for text context or Gemini).
    """
    # Logic similar to deepfake but for static images
    verdict = "VERIFIED"
    if req.media_url and ("fake" in req.media_url.lower() or "scam" in req.media_url.lower()):
        verdict = "FAKE"
    
    return {
        "verdict": verdict,
        "confidence": 0.98,
        "details": "QR Signature match: SUCCESS" if verdict == "VERIFIED" else "UI Manipulation Detected in transaction hash"
    }
