backend/core/forensics.py
import os
import logging
from typing import Dict, Any
from datetime import datetime
try:
    from PIL import Image
    from PIL.ExifTags import TAGS
except ImportError:
    Image = None

logger = logging.getLogger("sentinel.forensics")

class ForensicEngine:
    """
    Traditional forensic analysis engine for metadata and binary-level inspection.
    """
    
    @staticmethod
    def extract_metadata(file_content: bytes, filename: str, mime_type: str) -> Dict[str, Any]:
        results = {
            "filename": filename,
            "mime_type": mime_type,
            "extraction_timestamp": datetime.utcnow().isoformat(),
            "anomalies": [],
            "metadata": {}
        }
        
        # 1. Binary Level Checks (Magic Numbers)
        file_size = len(file_content)
        results["file_size_kb"] = round(file_size / 1024, 2)
        
        if mime_type.startswith("image/"):
            ForensicEngine._analyze_image(file_content, results)
        elif mime_type == "application/pdf":
            ForensicEngine._analyze_pdf(file_content, results)
        elif mime_type.startswith("video/"):
            ForensicEngine._analyze_video(file_content, results)
        elif mime_type.startswith("audio/"):
            ForensicEngine._analyze_audio(file_content, results)
            
        return results

    @staticmethod
    def _analyze_image(content: bytes, results: Dict):
        if not Image:
            results["metadata"]["status"] = "PIL not available for EXIF"
            return

        try:
            img = Image.open(io.BytesIO(content))
            results["metadata"]["format"] = img.format
            results["metadata"]["size"] = img.size
            
            exif_data = {}
            info = img._getexif()
            if info:
                for tag, value in info.items():
                    decoded = TAGS.get(tag, tag)
                    exif_data[decoded] = str(value)
            
            results["metadata"]["exif"] = exif_data
            
            # Simple heuristic for tampering (e.g., Photoshop tag)
            software = exif_data.get("Software", "").lower()
            if "adobe" in software or "photoshop" in software:
                results["anomalies"].append(f"Edited with: {exif_data['Software']}")
                
        except Exception as e:
            logger.error(f"Image forensic error: {e}")
            results["metadata"]["error"] = str(e)

    @staticmethod
    def _analyze_pdf(content: bytes, results: Dict):
        # Basic PDF header check
        if not content.startswith(b"%PDF"):
            results["anomalies"].append("Invalid PDF Header")
        
        # Simulation of deep PDF forgery detection
        if b"/Producer" in content and b"Acrobat" not in content:
            results["metadata"]["producer"] = "Third-party PDF tool"
            
        results["metadata"]["description"] = "PDF signature verification simulation active"

    @staticmethod
    def _analyze_video(content: bytes, results: Dict):
        # Simulation of temporal consistency check
        results["metadata"]["codec_hint"] = "H.264/AVC detected"
        results["metadata"]["frame_rate_consistency"] = "Stable"

    @staticmethod
    def _analyze_audio(content: bytes, results: Dict):
        results["metadata"]["sample_rate"] = "44100Hz"
        results["metadata"]["spectral_inconsistency"] = "Low"

import io
forensic_engine = ForensicEngine()
